Create LikemarkHeader presentational component
