+++ src/components/likemark-children/likemark-children.component.js
@@ -14,7 +14,7 @@ import LikemarkChild from '../likemark-child/likemark-child.component'
  * Display the Likemark Children as a list component.
  * @class @extends React.Component
  * @property {LikemarkChildState[]} props.children - The likemark children to display.
- * @property {function} props.onLikemarkClick - The function which trigger the state update when a likemark child is selected.
+ * @property {function} props.handleLikemarkClick - The function which trigger the state update when a likemark child is selected.
  */
 export default class LikemarkChildren extends React.Component {
   static get propTypes () {
@@ -23,17 +23,17 @@ export default class LikemarkChildren extends React.Component {
         id: PropTypes.number.isRequired,
         name: PropTypes.string.isRequired
       }).isRequired).isRequired,
-      onLikemarkClick: PropTypes.func.isRequired
+      handleLikemarkClick: PropTypes.func.isRequired
     }
   }
 
-  render () {
-    let { children, onLikemarkClick } = this.props
+  render = () => {
+    let { children, handleLikemarkClick } = this.props
 
     return (
       <ul>
         {children.map(child =>
-          <LikemarkChild key={child.id} name={child.name} onClick={() => onLikemarkClick(child.id)} />
+          <LikemarkChild key={child.id} name={child.name} onClick={() => handleLikemarkClick(child.id)} />
         )}
       </ul>
     )
